+++ database/migrations/2025_03_01_183347_create_tiktok_requests_table.php
@@ -10,18 +10,20 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('tiktok_requests', function (Blueprint $table) {
-            $table->ulid('id')->primary();
-            $table->string('action')->nullable();
-            $table->text('url')->nullable();
-            $table->json('request')->nullable();
-            $table->string('request_id')->nullable();
-            $table->string('code', 50)->nullable();
-            $table->string('message')->nullable();
-            $table->json('response')->nullable();
-            $table->string('error')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('tiktok_requests')) {
+            Schema::create('tiktok_requests', function (Blueprint $table) {
+                $table->ulid('id')->primary();
+                $table->string('action')->nullable();
+                $table->text('url')->nullable();
+                $table->json('request')->nullable();
+                $table->string('request_id')->nullable();
+                $table->string('code', 50)->nullable();
+                $table->string('message')->nullable();
+                $table->json('response')->nullable();
+                $table->string('error')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**